NSL-5191:  Change Password: working again
This seems to have stopped working when I upgraded the Rails 7.1 with more Turbo.

This is an urgent fix and is a bit rough around the edges, but has to go in pronto.

There were JavaScript errors showing up in the console and part of the fix was to replace
rendering with redirects.